Support full Unicode in lexer
