+++ juniper/src/parser/tests/mod.rs
@@ -1,3 +1,2 @@
 mod document;
-mod lexer;
 mod value;
